requirements.txt
# Azure SDK packages for Document Translation and Document Intelligence
# Note: Using specific versions for Python 3.13 compatibility
azure-ai-translation-document==1.1.0
azure-ai-formrecognizer==3.3.3
azure-storage-blob==12.23.1
azure-identity==1.19.0
azure-core==1.32.0

# Web UI Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
